Add files via upload
This project aims to develop a machine learning model to help diagnose and predict the severity of Polycystic Ovary Syndrome (PCOS). PCOS is a hormonal disorder that affects women of reproductive age and can cause a range of symptoms, including irregular periods, infertility, excessive hair growth, and acne. This project is implemented using python programming. The code and resources are uploaded.